Added validators to the user input
